re-organized contacts controller functions
